Arquivo car.py (models) modificado; Coluna 'daily_cost' da tabela cars atualizada para receber valores 'Float' em vez de 'Integer'.
Arquivo car.py (resource) modificado;
Argumento da coluna 'color' da tabela 'cars' não mais obrigatório e com valor default atualizado;
Classe BrandList transferida para arquivo criado carBrand.py (resource).
Método 'get' da classe CarList implementado;
Método 'delete' da classe CarResource implementado.

+++ code/models/car.py
@@ -9,7 +9,7 @@ class CarModel(db.Model):
     name = db.Column(db.String(50))
     color = db.Column(db.String(50))
     year = db.Column(db.Integer)
-    daily_cost = db.Column(db.Integer)
+    daily_cost = db.Column(db.Float(precision=2))
 
     brand_id = db.Column(db.Integer, db.ForeignKey('car_brands.id'))
     car_brand = db.relationship('CarBrand')
